stash. bring back useSitePlans
