+++ client/my-sites/plans-features-main/hooks/data-store/use-pricing-meta-for-grid-plans.ts
@@ -34,8 +34,7 @@ const usePricingMetaForGridPlans: UsePricingMetaForGridPlans = ( {
 		getSitePlanSlug( state, selectedSiteId )
 	);
 	const pricedAPIPlans = usePricedAPIPlans( { planSlugs: planSlugs } );
-	const pricedAPISitePlans = Plans.usePricedAPISitePlans( { siteId: selectedSiteId } );
-	const introOffers = Plans.useIntroOffers( { siteId: selectedSiteId } );
+	const sitePlans = Plans.useSitePlans( { siteId: selectedSiteId } );
 	const planPrices = useSelector( ( state: IAppState ) => {
 		return planSlugs.reduce( ( acc, planSlug ) => {
 			const availableForPurchase =
@@ -116,12 +115,15 @@ const usePricingMetaForGridPlans: UsePricingMetaForGridPlans = ( {
 	 * - For now a simple loader is shown until these are resolved
 	 * - We can optimise Error states in the UI / when everything gets ported into data-stores
 	 */
-	if ( pricedAPISitePlans.isFetching || ! pricedAPIPlans ) {
+	if ( sitePlans.isFetching || ! pricedAPIPlans ) {
 		return null;
 	}
 
 	return planSlugs.reduce( ( acc, planSlug ) => {
+		// pricedAPIPlans - should have a definition for all plans, being the main source of API data
 		const pricedAPIPlan = pricedAPIPlans[ planSlug ];
+		// pricedAPISitePlans - unclear if all plans are included
+		const sitePlan = sitePlans.data?.[ planSlug ];
 
 		return {
 			...acc,
@@ -130,7 +132,7 @@ const usePricingMetaForGridPlans: UsePricingMetaForGridPlans = ( {
 				discountedPrice: planPrices[ planSlug ]?.discountedPrice,
 				billingPeriod: pricedAPIPlan?.bill_period,
 				currencyCode: pricedAPIPlan?.currency_code,
-				introOffer: introOffers?.[ planSlug ],
+				introOffer: sitePlan?.introOffer,
 			},
 		};
 	}, {} as { [ planSlug: string ]: PricingMetaForGridPlan } );
