+++ packages/data-stores/src/plans/index.ts
@@ -19,7 +19,7 @@ export type {
 	PlanSimplifiedFeature,
 } from './types';
 
-export { default as usePricedAPISitePlans } from './queries/use-priced-api-site-plans';
+export { default as useSitePlans } from './queries/use-site-plans';
 export { default as useIntroOffers } from './hooks/use-intro-offers';
 
 // plansSlugs is a list with the identifiers for each plan and they are agnostic of billing period; eg: 'free', 'personal'
